chore: move Market to utils/type.ts
