+++ app/markets/components/MarketTableBody.tsx
@@ -4,11 +4,11 @@ import { ExternalLinkIcon } from '@radix-ui/react-icons';
 import Image from 'next/image';
 import { FaShieldAlt } from 'react-icons/fa';
 import { GoStarFill, GoStar } from 'react-icons/go';
-import { Market } from '@/hooks/useMarkets';
 import { formatReadable } from '@/utils/balance';
 import { getMarketURL } from '@/utils/external';
 import { getNetworkImg } from '@/utils/networks';
 import { findToken } from '@/utils/tokens';
+import { Market } from '@/utils/types';
 import { ExpandedMarketDetail } from './MarketRowDetail';
 import { TDAsset, TDTotalSupplyOrBorrow } from './MarketTableUtils';
 import { MarketAssetIndicator, MarketOracleIndicator, MarketDebtIndicator } from './RiskIndicator';
