+++ app/positions/components/RebalanceCart.tsx
@@ -9,7 +9,7 @@ import {
   Button,
 } from '@nextui-org/react';
 import { formatUnits } from 'viem';
-import { Market } from '@/hooks/useMarkets';
+import { Market } from '@/utils/types';
 import { GroupedPosition, RebalanceAction } from '@/utils/types';
 import { MarketBadge } from './MarketBadge';
