+++ src/hooks/useUserPositions.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { SupportedNetworks } from '@/utils/networks';
 import { MarketPosition, UserTransaction } from '@/utils/types';
 
@@ -127,79 +127,83 @@ const useUserPositions = (user: string | undefined) => {
   const [history, setHistory] = useState<UserTransaction[]>([]);
   const [error, setError] = useState<unknown | null>(null);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        setLoading(true);
-        const [responseMainnet, responseBase] = await Promise.all([
-          fetch('https://blue-api.morpho.org/graphql', {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-              query,
-              variables: {
-                address: user,
-                chainId: SupportedNetworks.Mainnet,
-              },
-            }),
-          }),
-          fetch('https://blue-api.morpho.org/graphql', {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-              query,
-              variables: {
-                address: user,
-                chainId: SupportedNetworks.Base,
-              },
-            }),
-          }),
-        ]);
-
-        const result1 = await responseMainnet.json();
-        const result2 = await responseBase.json();
-
-        const marketPositions: MarketPosition[] = [];
-        const transactions: UserTransaction[] = [];
-
-        for (const result of [result1, result2]) {
-          // eslint-disable-next-line @typescript-eslint/prefer-optional-chain
-          if (result.data && result.data.userByAddress) {
-            marketPositions.push(
-              ...(result.data.userByAddress.marketPositions as MarketPosition[]),
-            );
-
-            const parsableTxs = (
-              result.data.userByAddress.transactions as UserTransaction[]
-            ).filter((t) => t.data?.market);
-            transactions.push(...(parsableTxs as UserTransaction[]));
-          }
-        }
-
-        const filtered = marketPositions.filter(
-          (position: MarketPosition) => position.supplyShares.toString() !== '0',
-        );
-
-        setHistory(transactions);
-
-        setData(filtered);
-        setLoading(false);
-      } catch (_error) {
-        setError(_error);
-        setLoading(false);
-      }
-    };
-
+  const fetchData = useCallback(async () => {
     if (!user) return;
 
-    fetchData().catch(console.error);
+    try {
+      setLoading(true);
+      const [responseMainnet, responseBase] = await Promise.all([
+        fetch('https://blue-api.morpho.org/graphql', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            query,
+            variables: {
+              address: user,
+              chainId: SupportedNetworks.Mainnet,
+            },
+          }),
+        }),
+        fetch('https://blue-api.morpho.org/graphql', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            query,
+            variables: {
+              address: user,
+              chainId: SupportedNetworks.Base,
+            },
+          }),
+        }),
+      ]);
+
+      const result1 = await responseMainnet.json();
+      const result2 = await responseBase.json();
+
+      const marketPositions: MarketPosition[] = [];
+      const transactions: UserTransaction[] = [];
+
+      for (const result of [result1, result2]) {
+        // eslint-disable-next-line @typescript-eslint/prefer-optional-chain
+        if (result.data && result.data.userByAddress) {
+          marketPositions.push(
+            ...(result.data.userByAddress.marketPositions as MarketPosition[]),
+          );
+
+          const parsableTxs = (
+            result.data.userByAddress.transactions as UserTransaction[]
+          ).filter((t) => t.data?.market);
+          transactions.push(...(parsableTxs as UserTransaction[]));
+        }
+      }
+
+      const filtered = marketPositions.filter(
+        (position: MarketPosition) => position.supplyShares.toString() !== '0',
+      );
+
+      setHistory(transactions);
+
+      setData(filtered);
+      setLoading(false);
+    } catch (_error) {
+      setError(_error);
+      setLoading(false);
+    }
   }, [user]);
 
-  return { loading, data, history, error };
+  useEffect(() => {
+    fetchData().catch(console.error);
+  }, [fetchData]);
+
+  const refetch = useCallback(() => {
+    fetchData().catch(console.error);
+  }, [fetchData]);
+
+  return { loading, data, history, error, refetch };
 };
 
 export default useUserPositions;
